get LLVM JIT execution working for basic expressions

+++ src/core/ast/node/expression/integer.rs
@@ -3,5 +3,5 @@ use crate::util::source::Span;
 pub struct Integer {
     pub span: Span,
 
-    pub literal: isize,
+    pub literal: i64,
 }
